+++ packages/graph-explorer/src/components/Chip/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Chip";
-export type { ChipProps } from "./Chip";