+++ packages/graph-explorer/src/components/Tabular/ModuleContainerTabularHeader/ModuleContainerTabularHeader.styles.ts
@@ -6,7 +6,6 @@ const defaultStyles: ThemeStyleFn = ({ theme }) => {
   return css`
     &.entities-tabular-header {
       position: relative;
-      z-index: 100;
       display: flex;
       align-items: center;
       width: 100%;